The Infinite Terrain System.
A basic unit of terrain is a tile with size 256-by-256, as in the demo scene TrainAndTerrain.
InfiniteTerrainDriver.cs contains a few variable to tweak speed and acceleration.

The motion is a simple constant acceleration model, per GCSE physics.
At least 3 terrains are needed to support the illusion of unending train ride.